# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Octo Protocol API
  version: 1.0.0
  description: >
    OpenAPI contract for the /v1 Octo Protocol routes.


    Every `amount_stroops` field below is a JSON `integer` and stays that way in this version —
    that is unchanged by issue #215. Storage now also carries an arbitrary-precision
    `amount_base_units NUMERIC(78,0)` column internally (needed so a future, non-Stellar chain
    whose amounts don't fit 64 bits can be represented exactly), but nothing about that is exposed
    on these schemas yet. The public cutover to a string-typed amount field is a separate,
    versioned breaking change tracked in issue #227 — do not assume `amount_stroops` stays numeric
    across a future major version.
servers:
  - url: /
paths:
...
  /v1/wallets/{id}/withdraw:
    post:
      summary: Withdraw funds
      operationId: withdraw
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                destination:
                  type: string
                amount_stroops:
                  type: integer
                asset:
                  type: object
                  properties:
                    code:
                      type: string
                    issuer:
                      type: string
                  required:
                    - code
                    - issuer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
...
components:
  schemas:
...
    WithdrawResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - id
            - status
            - destination
            - amount_stroops
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
            stellar_tx_hash:
              type: string
            destination:
              type: string
            amount_stroops:
              type: integer
...